// src/app/modules/hcm/shared/models/holidays/calculate-holiday-cycle.ts
export class CalculateHolidayCycle{
    idCycleHolidays: number = -1;
    idLaborRelationship: number = -1;
    idStatus: number = -1;
    cycleHolidays: string = '';
    enjoymentDay: number = -1;
    additionalDays: number = -1;
    endDate: string = '';
    daysAvailable: number = 0;
    minDate: string ='';
}