// src/app/modules/hcm/shared/models/holidays/holiday-cycle.ts
export class HolidayCycle{
    enjoyDays: number = 0;
    additionalDays: number = 0;
    enjoyRealDays: number = 0;
    idStatus: number = -1;
    idTypeHoliday: number = -1;
    typeHoliday: string = '';
    idHolidayProgramation: number = -1;
    idCalendar: number = -1;
    idPayrollType:number = -1;
    starDate: string = '';
    endDate: string = '';
    payDate: string = '';
    status: string = '';
    idCycleHolidays: number = -1;
    cycleHolidays: string = '';
    dayProgramation: number = 0;
    daysAvailable: number = 0;
    minDate: string ='';
}